requirements.txt
fastapi==0.93.0
uvicorn==0.20.0
google-cloud-documentai==2.24.0
google-cloud-aiplatform==1.42.1